UPSTREAM: driver/uart: Introduce a way for mainboard to override the baudrate
The rationale is to allow the mainboard to override the default
baudrate for instance by sampling GPIOs at boot.

A new configuration option is available for mainboards to select
this behaviour. It will then have to define the function
get_uart_baudrate to return the computed baudrate.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7f67f1a0aed3a96c40a8a222880c6dac29a6890a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 235daa4bf6b6467b5df675dcfe5041b7f62eeae3
Original-Change-Id: I970ee788bf90b9e1a8c6ccdc5eee8029d9af0ecc
Original-Signed-off-by: Julien Viard de Galbert <jviarddegalbert@online.net>
Original-Reviewed-on: https://review.coreboot.org/23713
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/931063
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>