UPSTREAM: mb/google/poppy/variant/nautilus: Change SlowSlewRateForSa setting
If switch to VT2 on nautilus, screen flicker appears. We found if we
rollback the change of slew rate setting, then the flicker issue will
be gone: https://review.coreboot.org/c/coreboot/+/22588
But nautilus board needs slew rate tuning to reduce EE noise, so we
decided to change only SlowSlewRateForSa to 2 (Fast/8) instead of
rollback the whole change of the CL:22588. It can remove the flicker on
VT2.

BUG=b:71397040
BRANCH=master
TEST=emerge-nautilus coreboot

Change-Id: Ia05345bcb140f6787ff0143b0ba6e7c014a26ea6
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 3dd88f175d1863534a8bdb5c19d718becc31e6f3
Original-Change-Id: Id1d4bd8b1316c02c783de708ec4658e030193a26
Original-Signed-off-by: Seunghwan Kim <sh_.kim@samsung.com>
Original-Reviewed-on: https://review.coreboot.org/23877
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/943587