UPSTREAM: intel/common/block/scs: Add ability to send early CMD0, CMD1
In order to improve boot time with emmc, add ability to send CMD0
and CMD1 early in romstage. This way, by the time system boots to
payload, we are ready to continue with emmc setup and we don't need
to send CMD0 in payload again, and wait for card to reset and be ready.

BUG=b:78106689
TESTS = Boot to OS
        Force early_mmc_wake_hw() to return error, recover in payload
        Force an error in payload, make sure system can recover/boot

Change-Id: Ide5a8aebaf092071f4c7ecfe3c4f9d0eddc3fe38
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b7fe7a1a8e033706f39c0fded5901f0f1dce7cfb
Original-Change-Id: I3488b077bf5100a1e0f2c879fb1436105607d25e
Original-Signed-off-by: Bora Guvendik <bora.guvendik@intel.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/25068
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1627626
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>