UPSTREAM: mb/google/fizz: Enable VMX
We are enabling at the kernel level, but that is triggering an issue
where FSP expects it to be disabled so it forces a cold reboot on
every warm reboot, clearing the ramoops logs.  Enabling in BIOS so it
matches what the kernel expects.

This is the same change that were done for eve:
https://review.coreboot.org/#/c/22449/

BUG=None
BRANCH=None
TEST=echo PANIC >  /sys/kernel/debug/provoke-crash/DIRECT
     check for /dev/pstore/console-ramoops

Change-Id: Ic84c919fd691a067784e09b90463d15719eaff4f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f12bb7bcf2894eb226028bf960621ccfc680ac71
Original-Change-Id: Icd0bd01f5aee4c89f503eebba0808a1f3059e739
Original-Signed-off-by: Shelley Chen <shchen@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25251
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/970363
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>