UPSTREAM: soc/intel/common: Add function to check if xDCI is allowed
When CONFIG_VBOOT is enabled then the xDCI controller should only be
enabled if the system is in developer mode.  This prevents a system
in normal/verified mode from being used as a USB peripheral device
which could potentially be used to access user data.

This change adds a function to return whether xDCI can be enabled
or not, which will be used by the SOCs.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ifc4bd2e2efdebe1dbef09512cf5c34071a3f28a8
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 4df7d2c4953822c33be77e20e2ceff896e4a65c5
Original-Change-Id: Ie3ee9dd7077c094a01fd857a2e4033a12ce8979b
Original-Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25347
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/987667
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>