UPSTREAM: rockchip: Add gpio_set() function
The <gpio.h> API is supposed to include a gpio_set() function that just
toggles the state of a GPIO already configured as an output, even though
we rarely need it since gpio_output() can already be used to initialze a
GPIO to a default value while configuring it as an output. This function
was forgotten on Rockchip, so this patch adds it now.

BUG=none
BRANCH=none
TEST=none

Change-Id: Icea9b1d808643574f2b706331ef51bd8cbf5be96
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: ffeee420912eecf525564bb35b095b6bfa5d0de6
Original-Change-Id: I201288139a2870e71f55a7af0d79d4a460b30a0c
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25393
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/998065