UPSTREAM: spi: Add helper functions for bit-banging
Sometimes when bringing up a new board it can take a while until you
have all the peripheral drivers ready. For those cases it is nice to be
able to bitbang certain protocols so that you can already get further in
the boot flow while those drivers are still being worked on. We already
have this support for I2C, but it would be nice to have something for
SPI as well, since without SPI you're not going to boot very far.

This patch adds a couple of helper functions that platforms can use to
implement bit-banging SPI with minimal effort. It also adds a proof of
concept implementation using the RK3399.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5e50172a25f6cb82b12b21625011411ec6e80e5c
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 4783db2cf18c31ab48b219dabe1e04bf7f311d52
Original-Change-Id: Ie3551f51cc9a9f8bf3a47fd5cea6d9c064da8a62
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/25394
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/998066