UPSTREAM: drivers/aspeed/Kconfig: Select HAVE_VGA_TEXT_FRAMEBUFFER
This allows VGA output in SeaBIOS to be enabled using the
SEABIOS_VGA_COREBOOT Kconfig option. Currently, its impossible to select
the VGA text frame buffer in the Kconfig menu.

I'm not sure why this wasn't enabled in the first place, but
SeaVGABIOS seems to work just fine with this patch.

Tested on KCMA-D8.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6fe01c79e7e9958235818d3326241aa78dffe58f
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: e4dbd368aca5ef9c06835149bc6cba688a60519e
Original-Change-Id: Ic924a12fbe89940b5f26d211eb8de6cab0be767a
Original-Signed-off-by: Daniel Grber <dxld@darkboxed.org>
Original-Reviewed-on: https://review.coreboot.org/25554
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1007423