UPSTREAM: siemens/mc_apl1: Fix accuracy issue with IDT PMIC
Due to an accuracy issue on IMON in the IDT PMIC, the reported system
power consumption was higher than the actual consumption. To prevent
this problem, a logic must be implemented in mainboard_init(). This logic
consists of slope and offset as constants for Vcc and Vnn, which need
to be programmed by coreboot. This fix compensates for the accuracy
issue.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8ccf29e6418fbfdd7d68bf8fee39fe6df3779fef
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 0af272c1a9b0c9ca8338456a56c2bac587cc5051
Original-Change-Id: I77faf95951d03ac6ce97a6721dba6e8466122a25
Original-Signed-off-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/25585
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/1012430