UPSTREAM: nb/intel/i945: Put stage cache in TSEG
TSEG is not accessible in ring 0 after it is locked in ramstage, in
contrast with cbmem which remains accessible. Assuming SMM does not
touch the cache this is a good region to cache stages.

Tested on Intel D945GCLF and Lenovo Thinkpad X60, on cold boot the
external stage cache gets created and the stage cache gets properly
used on S3 resume.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iad763ef14c7d22bf888d4d24da08f202fbcbee77
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dce3927f208c75ec854f966e99c86a8081aca42d
Original-Change-Id: I447815bb0acf5f8e53834b74785d496f9d4df1da
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/25603
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1434488
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>