UPSTREAM: nb/intel/gm45: Put stage cache in TSEG
TSEG is not accessible in ring 0 after it is locked in ramstage, in
contrast with cbmem which remains accessible. Assuming SMM does not
touch the cache this is a good region to cache stages.

Tested on Lenovo thinkpad X200: on cold boot the external stage cache
gets created and the cached ramstage gets successfully used on the S3
resume path.

BUG=none
BRANCH=none
TEST=none

Change-Id: I72b6f21f220e4aa8467c6f6d40427c489faa711f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c3e9ba03b6e8f888680b0117df8d6405eebfd01a
Original-Change-Id: I642f7d6ae5523a35904c8e1f029027565a364d26
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/25604
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1434489
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>