UPSTREAM: nb/intel/x4x: Put stage cache in TSEG
TSEG is not accessible in ring 0 after it is locked in ramstage, in
contrast with cbmem which remains accessible. Assuming SMM does not
touch the cache this is a good region to cache stages.

Tested on Intel DG41WV, the stage cache gets properly created and used
on S3 resume.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5b925dfd199c54ce7b26e328ab2cb7105594f91a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a402a9e7ab4ce46bc8829646e59cffa079309590
Original-Change-Id: Ie46c1416f8042d5571339b36e1253c0cae0684b8
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/25606
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1434491
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>