UPSTREAM: nb/intel/sandybridge: support more XMP timings
Tested with a pair of GSkill F3-1866C9-8GSR.

This makes sure in particular that we honor the CMD rate requested by
the XMP profile. This memory kit needs a CMD rate of 2 to be stable at
DDR3-1600 and up, even though it passes training at 1.

Also respect requested CWL to match vendor firmware and for a potential
increase in performance. The tested kit requests a tighter value than
the per-frequency table provides and has shown to be stable using that
setting.

BUG=none
BRANCH=none
TEST=none

Change-Id: I43c02b74919aa98fccbac35a091c622dbaf9630d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dabebc3716bac1dbe9445dfb2bbbbb833ee21de1
Original-Change-Id: I634bed764d76345c27f02a2fae5abb2d81b38fd9
Original-Signed-off-by: Dan Elkouby <streetwalkermc@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/25664
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1014076
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>