UPSTREAM: soc/amd/common/block/pi/heapmanager.c: Simplify code
There are sections of code that are almost identical and they can be
converted to auxiliary procedures. For allocating heap, 3 sizes (the
buffer size of currently being examined node, the buffer size of the
current best fit node and the minimum size for a buffer that will need
to be split if selected as the best fit) are used often so they could
be stored in temporary variables. These 2 changes will make code shorter,
with less indentation problems and overall easier to read. The actual
logic of the code is not changed.

BUG=b:77940747
TEST=Build and boot grunt.

Change-Id: I34ed8694c3ffd600796f75aadf5b0ace8def6933
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 51895d183861234db9500bd1ee33634776e91e93
Original-Change-Id: Ib4c69981eab7452228ccae9ed9bc288c8baceffe
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/25703
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1021773
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>